requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
Django==2.2
djangorestframework==3.9.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.1
idna==2.8
more-itertools==7.0.0
oauthlib==3.0.1
pluggy==0.9.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytest==4.4.0
pytest-cov==2.6.1
pytest-django==3.4.8
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
soupsieve==1.9
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.1
